+++ lib/recording_studio_billing/billable.rb
@@ -9,7 +9,6 @@ module RecordingStudioBilling
     included do |base|
       RecordingStudioBilling.register_capabilities!
       RecordingStudio.enable_capability(:billing, on: base)
-      RecordingStudioAccessible::Compatibility.register_access_capability!
       RecordingStudio.enable_capability(:accessible, on: base)
     end
   end