+++ test/dummy/config/initializers/recording_studio_root_switchable.rb
@@ -29,8 +29,6 @@ RecordingStudioRootSwitchable.configure do |config|
         end
       end
     end
-    scope.access_check = ->(**) { true }
-
     scope.default_root = lambda do |roots:, **|
       roots.first
     end